backend/requirements.txt
# FastAPI LLM Backend - Phase 1 Requirements
# Basic FastAPI setup with no database or external services

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# Development tools (optional)
# pytest==7.4.3
# httpx==0.25.2  # For testing

# LLM provider libraries (optional)
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
tiktoken==0.5.2
httpx==0.26.0

# Database 
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Authentication
bcrypt==4.1.2
pyjwt==2.8.0
python-jose[cryptography]==3.3.0